[#11] Add Support For Fetching CryptoCurrency Prices
Add a getDailyCryptoPrices function to the AlphaVantage module for
fetching prices for a cryptocurrency.

Have the fetchPrices function take a list of cryptocurrencies in
addition to the CommoditySymbols from a ledger file. It will split the
commodities into equities & cryptos and hit the appropriate AlphaVantage
route to get the prices.

Add a `cryptocurrencies` configuration option and `-c`/`--crypto` flags
to the exe, allowing users to specify the cryptocurrencies to fetch
prices for.

Eventually, we should ditch manual specification & use AlphaVantage's
supported Digital Currency list:
https://www.alphavantage.co/digital_currency_list/

Refs #11