Fix AlphaVantage Crypto Route Breakage
Alphavantage has changed the `DIGITAL_CURRENCY_DAILY` endpoint so that
it returns data in the same format as the `TIME_SERIES_DAILY` endpoint.
We address this by removing the `CryptoPrices` type and have the
`CryptoPriceList` wrap a list of normal `Prices` values.

We change the `pVolume` field of the `Prices` type to parse into a
scientific number, so that we properly parse the decimal values returned
by the crypto endpoint. `parseScientific` has been changed to explicitly
parse a `Scientific` type instead of any type that implements `Read`.

AlphaVantage also changed it's error format, now returning messages in
an `Information` field instead of a `Note` field. We adapt by attempting
to parse error messages out of either field.